moves class substitution to own file
