+++ feature_engine/creation/combine_with_reference_feature.py
@@ -12,7 +12,6 @@ from feature_engine.dataframe_checks import (
     check_X,
 )
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _drop_original_docstring,
     _feature_names_in_docstring,
     _fit_not_learn_docstring,
@@ -20,6 +19,7 @@ from feature_engine._docstrings.docstrings import (
     _missing_values_docstring,
     _n_features_in_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.tags import _return_tags
 from feature_engine.variable_manipulation import _find_or_check_numerical_variables
