+++ feature_engine/creation/cyclical_features.py
@@ -6,7 +6,6 @@ from sklearn.utils.validation import check_is_fitted
 
 from feature_engine.base_transformers import BaseNumericalTransformer
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _drop_original_docstring,
     _feature_names_in_docstring,
     _fit_transform_docstring,
@@ -14,6 +13,7 @@ from feature_engine._docstrings.docstrings import (
     _variables_attribute_docstring,
     _variables_numerical_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.variable_manipulation import _check_input_parameter_variables
