+++ feature_engine/creation/math_features.py
@@ -5,7 +5,6 @@ from sklearn.utils.validation import check_is_fitted
 
 from feature_engine.creation.base_creation import BaseCreation
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _drop_original_docstring,
     _feature_names_in_docstring,
     _fit_not_learn_docstring,
@@ -13,6 +12,7 @@ from feature_engine._docstrings.docstrings import (
     _missing_values_docstring,
     _n_features_in_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 
 
 @Substitution(
