+++ feature_engine/discretisation/decision_tree.py
@@ -10,13 +10,13 @@ from sklearn.utils.multiclass import check_classification_targets, type_of_targe
 
 from feature_engine.base_transformers import BaseNumericalTransformer
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
     _n_features_in_docstring,
     _variables_attribute_docstring,
     _variables_numerical_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.tags import _return_tags
 from feature_engine.variable_manipulation import _check_input_parameter_variables
