+++ feature_engine/discretisation/equal_width.py
@@ -7,13 +7,13 @@ import pandas as pd
 
 from feature_engine.discretisation.base_discretiser import BaseDiscretiser
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
     _n_features_in_docstring,
     _variables_attribute_docstring,
     _variables_numerical_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.variable_manipulation import _check_input_parameter_variables
