+++ feature_engine/encoding/one_hot.py
@@ -8,12 +8,12 @@ import pandas as pd
 from sklearn.utils.validation import check_is_fitted
 
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
     _n_features_in_docstring,
     _variables_attribute_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.encoding._docstrings import (
     _ignore_format_docstring,
     _variables_docstring,
