+++ feature_engine/encoding/probability_ratio.py
@@ -7,13 +7,13 @@ import numpy as np
 import pandas as pd
 
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
     _inverse_transform_docstring,
     _n_features_in_docstring,
     _variables_attribute_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.encoding._docstrings import (
     _errors_docstring,
     _ignore_format_docstring,
