+++ feature_engine/encoding/rare_label.py
@@ -8,12 +8,12 @@ import numpy as np
 import pandas as pd
 
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
     _n_features_in_docstring,
     _variables_attribute_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.encoding._docstrings import (
     _ignore_format_docstring,
     _variables_docstring,
