+++ feature_engine/outliers/artbitrary.py
@@ -12,7 +12,6 @@ from feature_engine.dataframe_checks import (
     check_X,
 )
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _feature_names_in_docstring,
     _fit_not_learn_docstring,
     _fit_transform_docstring,
@@ -20,6 +19,7 @@ from feature_engine._docstrings.docstrings import (
     _n_features_in_docstring,
     _variables_attribute_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.outliers.base_outlier import BaseOutlier
 from feature_engine.parameter_checks import _define_numerical_dict
 from feature_engine.tags import _return_tags
