+++ feature_engine/selection/drop_correlated_features.py
@@ -8,11 +8,11 @@ from feature_engine.dataframe_checks import (
     check_X,
 )
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
     _n_features_in_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.selection._docstring import (
     _missing_values_docstring,
     _variables_attribute_docstring,
