+++ feature_engine/selection/drop_duplicate_features.py
@@ -4,11 +4,11 @@ import pandas as pd
 
 from feature_engine.dataframe_checks import _check_contains_na, check_X
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
     _n_features_in_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.selection._docstring import (
     _missing_values_docstring,
     _variables_all_docstring,
