+++ feature_engine/selection/recursive_feature_addition.py
@@ -2,11 +2,11 @@ import pandas as pd
 from sklearn.model_selection import cross_validate
 
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
     _n_features_in_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.selection._docstring import (
     _cv_docstring,
     _features_to_drop_docstring,
