+++ feature_engine/selection/shuffle_features.py
@@ -9,11 +9,11 @@ from sklearn.utils.validation import check_random_state
 
 from feature_engine.dataframe_checks import check_X
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
     _n_features_in_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.selection._docstring import (
     _cv_docstring,
     _estimator_docstring,
