+++ feature_engine/selection/single_feature_performance.py
@@ -6,11 +6,11 @@ from sklearn.model_selection import cross_validate
 
 from feature_engine.dataframe_checks import check_X
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
     _n_features_in_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.selection._docstring import (
     _cv_docstring,
     _estimator_docstring,
