+++ feature_engine/timeseries/forecasting/window_features.py
@@ -4,7 +4,6 @@ import pandas as pd
 from sklearn.utils.validation import check_is_fitted
 
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _drop_original_docstring,
     _feature_names_in_docstring,
     _fit_not_learn_docstring,
@@ -13,6 +12,7 @@ from feature_engine._docstrings.docstrings import (
     _n_features_in_docstring,
     _variables_numerical_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.timeseries.forecasting.base_forecast_transformers import (
     BaseForecastTransformer,
 )
