+++ feature_engine/transformation/boxcox.py
@@ -8,13 +8,13 @@ import scipy.stats as stats
 
 from feature_engine.base_transformers import BaseNumericalTransformer
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
     _n_features_in_docstring,
     _variables_attribute_docstring,
     _variables_numerical_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.tags import _return_tags
 from feature_engine.variable_manipulation import _check_input_parameter_variables
