+++ feature_engine/transformation/log.py
@@ -8,7 +8,6 @@ import pandas as pd
 
 from feature_engine.base_transformers import BaseNumericalTransformer
 from feature_engine._docstrings.docstrings import (
-    Substitution,
     _feature_names_in_docstring,
     _fit_not_learn_docstring,
     _fit_transform_docstring,
@@ -17,6 +16,7 @@ from feature_engine._docstrings.docstrings import (
     _variables_attribute_docstring,
     _variables_numerical_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.tags import _return_tags
 from feature_engine.variable_manipulation import _check_input_parameter_variables
